Use the Github API repo sorting.
The Github API can sort by pushed time, no need to do it manually.
http://developer.github.com/v3/repos/#list-user-repositories

This is both more efficient, Github does the sorting, but it also
fixes a bug.  Because the list of results are paginated, if you have
more than a page's worth of repositories it will only sort that page.
This results in a false view of what has been recently pushed.

For a good example, try my github account "schwern" with and
without this patch.